Refactor Providers Structure
