+++ src/providers/AppProviders.tsx
@@ -5,7 +5,7 @@ import { Toaster } from 'react-hot-toast';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { SessionProvider } from 'next-auth/react';
 
-import { getQueryClient } from '@/lib/queryClient';
+import { getQueryClient } from './QueryClientProvider';
 
 /**
  * Global providers wrapper for the LangRoute application.
@@ -14,7 +14,7 @@ import { getQueryClient } from '@/lib/queryClient';
  *
  * @param children - The app content to wrap with providers
  */
-export default function Providers({ children }: { children: React.ReactNode }) {
+export default function AppProviders({ children }: { children: React.ReactNode }) {
 	const queryClient = getQueryClient();
 
 	return (
