+++ src/providers/index.ts
@@ -1,0 +1,5 @@
+// Main app providers
+export { default as AppProviders } from './AppProviders';
+
+// Query client utilities
+export { createQueryClient, getQueryClient } from './QueryClientProvider';
